[ledger] Enforce key size restriction on the protocol level.
Previously, we relied on manual checks to make sure the provided keys
fit into our size limit. With FIDL2, it's now possible to enforce this
restriction on protocol level. This should help making sure the
restriction is respected in all methods, as well as making it more
transparent to the users of the API.

LE-445 #done

Change-Id: I7c53026ef66314cd4bf9d0c1c804b97bff603da6