[pkgfs] use readdir to avoid negative blob lookups
Negative blob lookups (where the blob does not exist) are relatively slow.
For large packages (containing thousands of files) this dominates the install
time. Currently the number of blobs in blobfs is typically low, so the
readdir cost is lower than the cost of doing the blobfs checks for these
cases. This change introduces a dumb heuristic for this case that is not
expected to scale well. This heuristic should be replaced by a new API in
blobstore in the future.

PKG-12 #done ?

Change-Id: Iae9f2ddc9d6cc86c616cf5f6b286e23d656c8cc2