[build] Compile arm64 userland with -ffixed-x18 too
We've declared x18 reserved in the public Fuchsia ABI.
Clang will automatically reserve it for --target=aarch64-fuchsia.
Before that toolchain arrives, and when using GCC, make sure
it's explicitly reserved via the compiler flag.  This is the
same flag already used for building the kernel, where x18 is
also reserved for separate reasons.

Change-Id: Iae77d15331047fd6f9ebc1730cf74c89d0322df9