[modular_tests] fix triggers, queue_persistence tests
The triggers test used to fail with a timeout.

Same problem as
https://fuchsia-review.googlesource.com/c/peridot/+/133360: The flow
token is held onto for too long in the second timeout branch of the
terminate call, and therefore the test runs into its own timeout.

I tried to solve this once and for all in AsyncHolder::Terminate(),
but didn't get it to work, so a stopgap for now.

Also fixes queue_persistence test.

Change-Id: Id932e45e6cc808704b884ca97d09833a45babd97