[modular_tests] Fix clipboard after FIDL2 transition.
Convert clipboard agent storage to use OperationQueue.

- The pointer to a fidl service was placed on the capture
  list of a callback of a method invocation on itself. This
  was like this before the FIDL2 conversion, but no longer
  works in FIDL2. With Operations, the fidl service pointer
  can be kept as a member.

- Operations also ensure that the Peek operation executes
  only after the previous Push has completed. Before it
  was very likely like that, but not as obvious.

Fix unchecked access to a SizedVmoTransportPtr that can be null.
It was obviously converted from something that cannot be null
in the FIDL2 conversion, but because of the situation above
that code would not get to run, hence the test previously
failed and did not crash.

Change-Id: I08f11b2f502bc9a5bdaf427e6d179e4fe68fd5c0