[tools][wlan][bluetooth] Use cipd to download firmware
Set up a garnet-wide cipd.ensure and script in tools/.
Use it for for wlan and bluetooth firmware.
Delete the download-firmware.sh scripts.
Set up a jiri hook to run cipd.

The vulkan sdk and fonts scripts are a little more involved, and can be
ported to cipd at a later time.

Change-Id: I1a853c9ff07ecbae33f18db6353cb107962d73ed