[arm64] Move Linux device tree parsing from kernel to the boot shim
The arm64 kernel now requires that x0 points to the bootdata on entry.
The boot shim parses the Linux device tree to find the bootdata
rather than the kernel doing this.
The boot shim is also now responsible for finding the kernel command line
and ram size in the device tree and passing this to the kernel via
additional bootdata records.

mexec and the hypervisor already pass bootdata to the kernel via x0,
so no changes were needed there.

Change-Id: I927930508566b5ae8f26acd053de7b4f040ee194