[serve-updates] make UI more like `fx boot`
serve-updates now runs in the foreground and reports on the state of the SSH
tunnel. It also makes a reasonable attempt make sure that amber-srv is still
running. It no longer leaks amber-srv, so updating amber-srv now works, as
does switching to a new build directory with `fx set`.

The ssh tunnel still leaks when this script is shutdown, as there's no
protocol over ssh to remove a forward after it's been setup except to tear
down the entire control socket (which would terminate any active shells).
Given that it's a remote forward, this should be mostly non-destructive, and
the leak case should not have any major negative effects for the most common
use case of a single unnamed device.

The output from amber-srv is also suppressed so as to not produce output
about the amber http requests.

Change-Id: I94c25208d0a19071ba66a6ec67ff11b6d2cd3aca