[block][ahci] Remove barrier flags from driver
BLOCK_FL_BARRIER_BEFORE and BLOCK_FL_BARRIER_AFTER are
currently implemented in the block midlayer: if an incoming
operation has these barriers set, they will be used to
determine when operations should be enqueued to the underlying
driver. This logic is managed in dev/block/block/server.cpp,
and as a consequence, underlying drivers can process incoming
operations in any desired order.

Change-Id: I659dcc35296bef20ee7da0ea862db125b49b0095