[scenic][client] Flush when getting close to zx::channel limit
Previously, scenic would accumulate an unbounded number of commands on
the client and attempt to send them all in a single zx::channel::write
call. However, zx::channel::write has a finite capacity and would reject
the write.

Now, we send commands in chunks approximately the size of the channel
write capacity.

Change-Id: I40274e19d17518b60308674ba3aed80cd6dec3e2