[mexec] Fix Boot EL restore bug in mexec
Summary:
Fixes ZX-1945 where we observed that gcc/clang
builds of Zircon could not Zedboot into one
another.

Root Cause:
Commit bdd202a introduced a feature where we
attempt to mexec into the same EL that we
initially booted the system to in order to
facilitate hypervisor development.

For example, if the system initially booted
into EL2, we'll attempt to reboot the system
into EL2 also.

This requires us to call a shim in the EL2
which is resident in Zircon. Previously we
were calling this shim _after_ replacing the
existing kernel. This _only_ worked because the
location of the shim was mostly stable between
kernels (especially if they were generated) by
the same toolchain.

Problems obviously occurred when clang and gcc
put the shim at different locations.

Fix:
We now call this shim before supplanting the
existing kernel.

Validation:
Tested on Vim2 and Gauss by repeatedly
netbooting back and forth between two kernels
generated by clang and gcc respectively.
Note that these two kernels were not able to
mexec into one another prior to this fix.

ZX-1945 #done

Change-Id: I2e69c4fdcff40ea7db70287bf6b11be99f4ef9b9