[zircon][sdhci][amlogic]: Initialize emmc on vim2
This changeset does the EMMC initialization on vim2 and extracts the
EXT_CSD register. It fails to switch to high speed timing.
There are some missing parts(tuning, dma descriptors, asynchronous I/O).
I will add each of the missing functionalities in later changesets.
This changeset will be a no-op as the code that recognizes the emmc is
commented out.

Change-Id: I8c1371bad7f9bb2d09862d2ce148851df674f720