[ledger] Do not delay binding a just created page on the cloud.
Until now, the Ledger would delay binding a new created page until it
receives a response from the Cloud.
When the page is created locally, without an id, the response from the
cloud will always be empty. This CL ensures that the Ledger doesn't wait
on the Cloud in this situation.

LE-453 #done

Change-Id: Ib71bc7204a54b84cd30d36fc715ed1c7ff650c24