[media] run the in-proc framework single-threaded
This CL makes the in-proc streaming framework run on a single thread.
Individual nodes in the graph may still use additional threads, but
the framework itself uses only one.

There are still a lot of mutexes and atomics in the code that are no
longer needed. Those will be removed in the future.

Change-Id: I26aa9306d330f499d7d173225482ab1916fe70c4