[fdio] map ZX_ERR_CANCELED to EBADF
ZX_ERR_CANCELED means the handle corresponding to a fd is closed
by somebody while the caller is still waiting on it. POSIX doesn't
define how the event-wait APIs should behave on such a case (it is
basically a programming error). But probably EBADF is a better error
code than ECANCELED, which is currently returning.

ZX-1892

Change-Id: Ie6ca36991a6a401897757b6fa63903e9ecf8907e