[audio][mixer] AudioRenderer float support (emulated)
This CL enables the AudioRenderer APIs to accept audio data in the
float32 format. The underlying audio mixer core may or may not yet
natively support float32, but it can accept data in this format,
so with this CL any clients of the AudioRenderer APIs (v2 or v1)
should be able to provide data in that format.

MTWN-44 #done

Change-Id: I900a925c312039cb34fd17a21201f01b9f31b88d