[audio][mixer] Output device float32 support
If a device indicates support for AUDIO_SAMPLE_FORMAT_32BIT_FLOAT,
we should allow this to be selected as the device format, if no
better format is available.

Because we do not yet have hardware with which to verify this,
this CL keeps the priority of float LOWER (less preferred) than
16-bit-signed-int. When we have suitable hardware in hand, we can
test it with custom binaries before switching the default behavior.

MTWN-110 #done

Change-Id: I8e0bbe7c1690e7cd9229864f17b4599bd7bdb56f