[audio][mixer] MediaClient pass audio in float32
The MediaClient API currently accepts audio data in float32,
but converts it to int16 before passing it to the audio core.
It does this because the AudioRenderer API did not previously
support the float32 format natively. Now that AudioRenderer
does accept audio data in this format, this CL updates the
MediaClient API to pass along the audio data that it has
been provided, as-is without alteration.

MTWN-102 #done

Change-Id: I07aa83012af74fe38804c091fa2f88604f4c3097