[debugger] Frontend support for frames.
Supports listing and selecting frames in the console frontend.

Supports maintaining the frame list while the thread is stopped on the
client Thread object, and requesting the list on demand.

Plumbing to hook frame selection to the stack unwinder.

Some refactoring of the command context filling-out in the
ConsoleContext since adding frames made it too long.

Change-Id: Ib69964847fd1333c8e19cb6e4294eaae1805879e