[syscalls][vmo] Make zx_vmo_create_contiguous() accept BTI handles
Contiguous buffers are used to satisfy requirements of DMA engines,
so BTIs are a natural control handle for this.  This also opens up
the opportunity of modifying the zx_vmo_create_contiguous() contract
to let the kernel provide non-contigous physical memory that a DMA
remapping unit could make look contigous.

ZX-1813 #done

Change-Id: Ia0e42d7a6c2ef0108765b07fef3a58cf8cac0e0a