[async][ledger] Add a temporary CreateThread function.
Following armansito@ in
https://fuchsia-review.googlesource.com/c/garnet/+/141355, we create a
CreateThread variant that also returns an async_t*. This will be used to
initialize the only remaining Environment variable in ledger *not*
initialized with an async_t*. Once that is in place, we may begin to
safely remove the dependency on Evironment's main_runner(). Once that
is done, CreateThread will be deleted.

BUG: ZX-1819
Change-Id: If52ea7bddc975b812c46c6a651db0795f9e59a7f