[media] tracing tweak
This CL changes the way decode events are recorded in the trace log
so that the distinction between audio and video decoding is easier
to see in event visualization. As it was, differentiating between
an audio decode and a video decode required clicking on the event.
With this change, the two types of decode will have different names
and colors.

Change-Id: Iae4c6cb070a154d519e4c27f1d3aad803dece4d5