[make] Link against static libraries using --start-group/--end-group
Before this change, ordering of libraries in MODULE_STATIC_LIBS is
significant in the GCC build: targets (such as minfs) have to order
the libraries according to dependencies between the libraries.
However, that apparently wasn't the case in the Clang build (which
uses lld).

Fix this by using --start-group/--end-group, which tells the linker to
resolve dependencies between libraries in any order.  This makes the
GCC build behave like the Clang build in this respect.

The entries in minfs's MODULE_STATIC_LIBS can now be sorted.

Change-Id: If784b2f3e781721629f9c1fb3a6d57a95cde1354