[modular][daisy] Make 'null' a valid noun name for Daisies.
This null noun becomes a Link that *looks like* the default link to the
child module if it uses the GetLink() call. This eases the transition
away from default links by allowing engineers who are updating mods to
split up the work of transition a parent and all its children at the
same time.

MI4-749 #comment [modular][daisy] Make 'null' a valid noun name for Daisies.

Change-Id: I6218d3779c49c336eaa4b7854ba0b979dc87d870