[mesh] Mesh routing prototype
This is work in progress, and right now more of an opportunity for me to
learn Rust than anything useful for others.

There are three layers:
- mesh-protocol - protocol level pieces that do not encode any behavior
- mesh-router - routing pieces that can operate untrusted (actions that
don't rely on secrecy)
- mesh-endpoint - secure pieces implementing reliability and message
ordering

The Fuchsia implementation will be composed of two binaries: one that
runs mesh-router as a global singleton process, and a second that embeds
mesh-endpoint and provides privacy bits - with a FIDL interface between
them.

The Linux implementation will merge router & endpoint into a single
library.

Change-Id: I0bb725a1fce295b0d6ecbb113294636c7a8bcf3e