[fidl][go] Flesh out validation errors
This change fleshes out validation errors for the FIDL golang bindings.
It doesn't represent the complete set of possible errors, but serves to
help consolidate errors at the very least.

An important aspect of this change is that it removes any transitive
dependency on the "os" package, which will allow the bindings package to
be used from syscall/zx/fdio.

Change-Id: I5794f08b8d7a15eeb538301da72d40a2beb79ba7