[async] Strengthen invariants for list of pending waits.
Ensure that we don't try to cancel waits which are not in the list
because they are being dispatched concurrently or because the
loop is shutting down (because we don't drain the port packets).

Also be more strict about the set of errors returned.  In particular,
never return ZX_ERR_BAD_HANDLE since it indicated something is
deeply wrong with the state of the loop or the wait objects and tends
to balloon into hard to find bugs much further downstream (such
as during loop shutdown).

Change-Id: I67567a1e14bfbeaca0ff0bee9cb46faacdb53ba6