[ledger] Update serialization helpers for FIDL2.
FIDL2 introduced a new encoding format for messages. This change updates
serialization size computation helpers for Ledger, which rely on the
encoding layout details. This can be considered an intermediate solution
while we don't have a way to do pagination that wouldn't rely on FIDL
internals.

It also adds some tests to ensure our computation is correct.

LE-427 #done
LE-449 #comment The code in question is rewritten for FIDL2 as an
intermediate solution.

Change-Id: I2bb507a92c5c68dddf2fb0057193d3cac2c1f929