[uboringssl] Align uboringssl with the //third_party/boringssl
This CL moves the assembly files and updates rules.mk to be more
consistent with //third_party/boringssl's layout.  This allows the
scripts to check the subset and generate build files to be removed;
these are handled by roll_boringssl.go.  The README has similarly been
simplified.

ZX-1991 #comment Adds the necessary changes

Change-Id: Ifc09dd27cd90c5c5381182fcc4f27e5400b97d05