[usb][camera] Round up buffers to page boundaries
Vmars must be mapped on page boundaries, so having a buffer
with a fractional page size causes errors when I try to map
individual buffers.  This is fixed by rounding up the
max_frame_size, so each buffer will be padded.  This is not
very wasteful except where image size is very small:
160x120 images:
1 byte/pixel: 4.6875 pages, rounding gives 6.25% waste
2 byte/pixel: 9.375 pages, rounding gives 6.25% waste.
640x480 images (and most common larger formats) divide evenly
into page boundaries, so this patch has no effect on the more
commonly used sizes.

Change-Id: Ia18c74039f0fc5af959f372101ee4b36c41c4701