[fidl] Fix potential overflow issue in fidl encoding / decoding / validation.
The number of elements in a vector is user controlled and could result in overflow
when calculating the total size of the vector. In most cases this would cause another error
later on in encoding / decoding / validation, but I think it's possible that it'd result in
reading or writing past the end of the |bytes| buffer and thus memory corruption. In any
case this will at least provide a useful error message.

I believe this is the only remaining operation that uses untrusted inputs and could
result in overflow and incorrect behavior. Everything else relies on the fidl_type_t
structure being valid, but that is a trusted input since it's generated by the FIDL
compiler.

Added unit tests that fail before the fix and pass after it.

SEC-41 # comment: !

Change-Id: Ia48efca12eb9791eef14023b727429dcd2024875