[abigen] Allow array size to be a product of arguments
We would like to change fifo_write/read syscalls to take the number of
entries instead of the number of bytes.

Per discussion with Roland, taking just the number of entries as an
argument is not a good idea since it would be impossible to compute the
byte size of the buffer purely from the function's arguments.

As a compromise, in addition to the number of entries, we can also take
the size of an entry (and validate that it matches the value passed into
fifo_create()). That way, the total byte size of the buffer could be
computed as the product of the two.

Currently, the abigen language is not able to express this concept. The
goal is to have a syscall like this:

    syscall fifo_read(handle: zx_handle_t,
                      data: any[count * elem_size] OUT,
                      count: size_t,
                      elem_size: size_t)
        returns (zx_status_t, num_written: uint32_t);

Supporting an arbitrary arithmetic expression inside '[]' would probably
be an overkill for now. This change only adds support for specifying the
array size as a product of identifiers.

ZX-1951

Change-Id: Ie5cef46cdb631c92e2eaf0d163373030baa835e9