[auth] Token Manager interface changes.
This cl splits the token manager interface in such a way that:
1. The UI presentation layer could only be controlled by the component
that initializes token manager instance, and
2. The data layer that mints tokens will never have control of the UI
elements.

This helps us in preventing phishing attacks where user is prompted to
authorize on a site controlled by third party agents.

One more change that gets rid of hardcoded auth_provider enums and
replaced by an "url" is coming in a separate cl.

Change-Id: I01a7023bdde2cd74b03ffe53da682b150ec8f0a5