[media][tests] Additional resampler tests
This CL adds resampler tests for the Reset method and the
filter width properties (both positive and negative), for
PointSampler and LinearSampler. It also adds a number of
clarifying comments and DCHECKs to the associated resampler
code.

This test support will aid the development of additional
resamplers that might be added in the future.

change-Id: Ib0e6714b18bea364bc87cf241d58b448f929dae9