[trace] Prevent use-after-free when trace arg is a temporary.
This fixes an issue with temporary strings passed as arguments to
the trace macros.  When writing a record which has a string argument,
the string's data pointer is placed in a table of trace_arg_t
values which are then passed to the function which actually does
the writing.  That function will copy the contents of the string
as required.  However, if the string was a temporary, it may have
already been destroyed by the time that happens.

This change first assigns the trace arguments to local variables
to ensure they remain in scope for the whole block.

TRACE_DURATION_BEGIN("foo", "name", "key", fbl::String("string"));

Tested using ASAN that the problem has been fixed.

Change-Id: I97cb382242d7736bb9f0fa6e5c9f2ae64009ba38