[port] Fix name of count parameter in wait/queue
The port::wait/queue methods previously took a 'size' parameter. This
was passed to zx_port_{wait,queue} as the 'count' parameter, which led
to at least one instance of confusion. Rename the 'size' parameter to
'count' to reduce the potential for confusion.

Also fixes a caller that treated the 'size' parameter like a size.

Change-Id: Ic8dd5752c3004977b47e978d85c06c0d56f0443d