[usb][audio] Add classes to manage units and terminals.
Add classes to help to parse the descriptors for USB audio "units" and
"terminals", the building blocks of the paths for audio in a USB audio
device.  The blocks exist as part of the USB audio control interface
and are typically (but not always) bound to USB audio streaming
interfaces.

Change-Id: Ic13405eaae5891ef225c063771c36108057c78e0