[bt][examples] LE Heart Rate peripheral example
A brief example for the author to learn FIDL by consuming the C++
bindings to the GATT and BLE public APIs.

The Fuchsia device will advertise a GATT Heart Rate service whose heart
rate measurements will contain a "rate" as the % CPU time busy and an
"energy expended" as the number of context switches.

NET-712 #done

Test: Manual hardware testing against an Android central
Change-Id: I2457df11e120ca9d52debbb7bc061a088c3c71af