[usb][audio] Add a class to manage a USB audio streaming interface
Introduce a class which manages the details of discovering and parsing
all of the details of a USB audio streaming interface and each of its
alternate variations.  Eventually, this interface needs to be linked
to an AudioPath object in order to manage the flow of audio from the
host DMA through the terminals an units which make up the path through
the control interface hardware.

Change-Id: I219e78377dff0e7ff1a2aa9f3a5394da85f3fd5c