[kernel][thread] Add "resched_disable" count for deferring thread rescheduling
This is similar to "preempt_disable", except that it does not block
pre-emption by interrupt handlers.

We now have a single field in thread_t containing two disable counts
so that it can be checked with a single memory access.

The logic wrapping interrupt handling becomes a bit more complex.  I
have moved it into functions defined in a new header, interrupt.h, so
that it can be shared across architectures and tested by unit tests.
We now save and restore preempt_pending so that we can tell whether an
interrupt handler has set it to true.

Unit tests: For each test of preempt_disable that runs outside of
interrupts, I have added a test of resched_disable.

ZX-1690

Change-Id: I4d33d5c604709a31f3c9f2d82a18e5d24ad8b5f2