[fidl][go] Refactor BindingSet binding retrieval
This change refactors the BindingSet to allow retrieval of information
about bindings by returning a key. Currently there are no methods to
index into the BindingSet, but this is necessary for events so that an
event proxy can be created from a binding's channel.

It also adds support for a callback for when a binding is closed, so the
service may do whatever teardown it might need to.

Finally, this change updates all calls to BindingSet.Add to reflect the
new interface.

FIDL-97

Change-Id: Ibbf9c521aab6b5a4aa36807cd0588b43ea9c570a