[netstack][tests] Add NetStreamTest.BlockAcceptWriteNoClose
The test simulates an unexpected process exit so that no
Close op is sent for the listening socket. The handle associated
with the socket fd is closed so netstack should detect it
with a PEER_CLOSED signal and tear down the listening port
correctly. Otherwise the second bind on the same port would fail.

NET-694

Change-Id: I9c64856324be9bec9262b0f0f684194d8d670d5f