[fidl] MessageBuffer: Avoid possible overflow and check malloc ret val.
We promote the two 32-bit operands to 64 bits early in order to avoid overflow.
The result was already being promoted to 64 bits.

When malloc returns NULL, failing with a specific error makes debugging easier
(vs seeing a segmentation fault later on when the buffer is used).

Part of the security review of core FIDL 2 code.

SEC-41 # comment: done

Change-Id: Ic9d2c8ebafa111e427ac7f1e6fc923ad4fcd426b