[media][tests] Tests to target accumulated resampler error
Previous resampling tests performed the entire mix job with a single
call to produce the entire set of 64k frames. This is different from
how the Mixer objects are used by surrounding code in audio_server.
To better model this, we now call the Mixer objects with a sequence
of smaller requests. This is configurable, but the current setting
is buffers of approx. 5 ms.

The adjusted tests clearly demonstrate the effect of MTWN-49 and will
quantify the sonic improvement when this defect is fixed.

Because of the change in how we measure frequency response and sinad,
we have updated our baseline audio thresholds. For a brief time,
purely for the benefit of those curious to compare, the previous
values are also included in audio_result.cc (commented out), but these
will be removed in short order by subsequent CL.

MTWN-49 #comment

Change-Id: I7fff92d37d37c40856ee4f78b806381fac6b5cef