[ramdisk] Improve sleep_deferred test
This CL improves the unit test for ramdisk's "stall" feature.  Since
block_watcher and others can send unexpected txns to the ramdisk, Care
must be taken to precisely order starting a thread to wake the ramdisk,
scheduling a sleep for the ramdisk, and sending txns that are expected
to block.

This CL corrects the unit test, and adds comments to help other
consumers.

ZX-2001 #comment Fix block_watcher hang

Change-Id: Ia075d85fb8af5f315651bb5e99b106941bb40380