[escher] New shader compiler framework.  Details follow...
The main design goals are:

1) To decouple GLSL compilation from the production of VkShaderModules
(which are thin wrappers around binary SPIR-V code, used to generate
VkPipelines). It is quite likely that we will move away from runtime
GLSL compilation, and instead load pre-compiled/optimized SPIR-V.

2) To support "hot reload" functionality, allowing us to edit Scenic
shaders without needing to restart the service and its clients,
resulting in faster iterations.

The first goal is achieved by the ShaderModule class, which is a thin
wrapper around VkShaderModule that notifies dependants that any
VkPipelines built using the module must be rebuilt.

ShaderModule is designed to be subclassed.  The second goal is supported
by the ShaderModuleTemplate class, which acts as a factory to obtain
ShaderModules that are specialized by providing different macro
definitions to the GLSL preprocessor.

The preprocessor supports #include directives, and any files so included
are added to a watch-list, so that when a file changes, all template
variants that use it are recompiled.

Note: there is no actual filesystem-watching yet; this is mocked by a
placeholder HackFilesystem and HackFilesystemWatcher classes.  The goal
was to make it clear where to plug in a Linux/Fuchsia filesystem
abstraction, when it exists.

SCN-662 #done

Change-Id: If845a99398282a356edfd89b0f40860b5e8be080