[kernel][threads] Test suspending during exception handling.
Fixes and enables the test for suspending threads during exception
handling.

Moves sending the exception report to the port out of the
ThreadDispatcher lock. A reference to the port is held on the stack and
this potentially slow operation is not required to be in the lock.

Kills the process instead of trying the next exception handler when the
desired exception handler fails for some reason. Otherwise downstream
exception handlers might get an exception they aren't expecting.

Change-Id: Id3d96f09b01d138928180f104b493bbfa5b78ccb