[netconnector] fix call to service from utility
This CL fixes a problem with the way netconnector, running as a
utility, calls netconnector the service.

With the new implementation of fidl::InterfacePtr, the object
returned by dereferencing an InterfacePtr is part of the InterfacePtr
itself (specifically, operator-> returns &proxy_). This breaks a
calling pattern that used to work, namely:

    fidl::InterfacePtr<Foo> foo = ...
    foo->Bar([foo = std::move(foo)]() {
        ...
    });

This is intended to extend the lifetime of the channel through the
callback. Here, the -> operator runs (returning &foo->proxy_) before
the capture block is evaluated. Unfortunately, the std::move call
invalidates foo->proxy_ so that the call to foo->proxy_->Bar never
happens (does nothing).

The fix here is to call the service in synchronous mode.

NET-689 #done

Change-Id: I9e106cb884cb4d29b43c1fa5471a79d7efaca9b3