[debugger] Start on remote debugger.
Adds basic socket support for debug_agent. This is not used yet since
the client side is non-functional.

Adds a host MessageLoop that works on Linux (and theoretically Mac but I
haven't tested it yet). Fills out the main implementation for the host
build of the debugger such that the command line input works.

Updates LineInput to work on Linux terminals. This means putting the
terminal into "raw" mode during line input, and disabling raw mode when
printing normal output. The Fuchsia terminal is not as complete so
doesn't have or need this logic.

Adds stubs of commands to connect to a remote system in the client. Some
parsing for IP addresses and ports is added. The connection code is not
currently hooked up. The client socket implementation will be done in a
followup.

Change-Id: I237cf520bb9d9ab450075ef4e11fe12edf180480