[scripts] Use `awk` for AWK, use least-common-denominator features
Which implementation is installed varies between systems (e.g.  Linux
systems often have GNU awk or mawk, and nobody is entirely sure what
the macOS awk is).  But there should always be one that is installed
under the plain `awk` name, so use that.  The relocation processor has
to use only the standard awk features, which do not include parsing
hexadecimal.

Change-Id: I948c135117baa6991bfe5f89520acad1c6f31c82