[story_shell] Keep module manifest in story record.
The module manifest contains information needed by the story shell
to layout. In order to resume a story, we need this information,
so we need to record it.

Alternative ways to obtain the module manifest when resuming a story
are:

1. Looking up the manifest again from the package.

2. Resolving the module from the intent again.

We will need to do all of them in order to allow manifest changes to be
picked up by restarting a story during development, and to notice
manifest updates when requesting a module again, and restart an existing
module instance. In this CL, we just add a way to keep a module manifest
in the story record. Using the manifest will be a followup CL.

MS-1473 #comment Keep module manifest in story record.

Change-Id: I460c2865a69a8844420651ada91689b5959d787a