[story_shell] Resume modules with intents in StoryStartCall
Sending modules to the story shell requires the module manifest,
not just when a module is added to story first, but also when the
story is resumed.

The manifest could be recorded in the story record (upcoming in
separate change) along with the daisy that is already recorded,
but even if it is, it may have changed (happens in the code lab),
so we need to re-resolve it from the recorded intent.

MS-1473 #comment Resume modules with intents in StoryStartCall.

Change-Id: Ia961c49c8c3efdd2ae9bf485e530f6161b3593c7