[modular] Add modular::Future class.
This change also updates StoryControllerImpl to use Futures. This is
intended as both an example of how Futures are useful and to simplify
the StoryControllerImpl code.

Also add some small tests for the existing FutureValue class (which will
be replaced with the Future class once this is deemed ready).

TESTED: run_modular_tests.sh passes, unit tests pass, normal device
usage works.

Change-Id: I57f5a3d96f84aa86055e9d8f92513db1bff5f025