[threads] Fix threads test.
This test was causing a page fault (on purpose). It resumed from the
exception with the intent of stepping over it, but this actually just
re-executes the same instruction. I think the crashlogger was doing
something such that the test still worked, but it failed in userboot
mode.

This replaces the instruction with an explicit debug breakpoint and then
steps over it correctly.

ZX-2021

Change-Id: I72b214cfdcc6113da0a42b87c24c532d253baea7