[audio][mixer] Fix bad recycle impl in PendingFlushToken
Fix an issue in the implementation of the recycle handler for
PendingFlushToken which was double-free'ing the the object.

The idea here was supposed to be that when a flush operation was
started for a renderer, a flush token was created which owned the
flush complete callback and offered to each link that the renderer was
feeding at the moment.

If the other end of that link was actively consuming data at the
instant the flush happend, then the link would hold a reference to the
pending flush token, otherwise the link would release all of its
references to pending packets and not hold onto the flush token.

Eventually, all of the references to the pending flush token would go
away, and the flush token would become recycled.  Since this was the
first time the token had reached end of life, it would be ressurected
by the recycler, wrapped in a unique_ptr and posted to the main server
thread to invoke the callback.  Once the callback was finally invoked
by the main server thread, the unique ptr reference would go away and
the token would die again, this time actually being deleted.

The bug, howeve, was that the token was being unconditionally deleted
in the recycler, both in its first and second lives.  The fix is
simple, only delete the token if it had been recycled at least once.

Change-Id: Ibc3fe9c49eaa2f1d963ae77eb509952e6c5b2fc1