[zircon][xhci] Initialize TT fields for all the low/full speed devices
The transaction translator fields should be set for all the low/full
speed devices. Currently our code only sets it if the parent is a high
speed hub and does not pass on the TT fields to children of a full speed
or a low speed hub which is connected to a high speed hub. Without this,
the xhc generates errors for the SET_ADDRESS command and the device does
not initialize properly. This changeset does the following:
1. Passes on the TT fields as required.
2. Add a sleep between Address Device and Get Descriptor commands.
Without this, the Get Descriptor command fails on Razer Megalogon headset.
3. Provides error handling for legacy devices that might expect Get
Descriptor request before SET_ADDRESS request.
ZX-1933 #done ZX-1975 #done

Change-Id: I80045a2569ba108f697034dd8c85229a8960bea9