[amber] Implement API to wait for package readiness
Implement the GetUpdateComplete API which does not return until
the package is activated after the package's components are
available.

This functionality is primarily implemented by ActivationMonitor
which receives requests to install packages. It writes the meta
FAR to pkgfs and then monitors for the signal that the package is
activated.

F5-32 #done

Change-Id: I3c090f07799329c24d335dc3ae209ba744867cfe