[escher] Add a generic ObjectPool.
This pool continguously allocates objects from blocks of memory.
It supports both constructing/destroying objects as they are
allocated/freed, and also when whole blocks of memory are
allocated/freed.  The latter approach is useful for managing
Vulkan resources such as descriptor sets, which must be allocated
in batches.

SCN-691 #done

Change-Id: I7a0a80982e251b4d822c128a3d623e28ebff16b8